Store source location in a single object.
Source file, function name and line number are now stored in a const
static container object. This has the following benefits:
- Slightly lighter profiling workload (3 instructions less).
- Profiling queue event size is significantly reduced, by 12 bytes. This
  has an effect on all queue event types.
- Source location grouping has now no cost, as it's performed at the
  compilation stage. This allows simplification of server code.
The downside is that the full source location resolution is now
performed in two steps, as the server has to query both source location
container and strings contained within. This has almost no real impact
on profiler operation.

+++ server/TracyEvent.hpp
@@ -10,7 +10,7 @@ struct Event
 {
     int64_t start;
     int64_t end;
-    uint32_t srcloc;
+    uint64_t srcloc;
     uint32_t color;
 
     Event* parent;
